Transform correlation

Transform correlation

Usage

transform_rho(rho, el)

Arguments

rho

correlation matrix

el

a K-by-3 matrix in which the first two columns are row and column indicators; and the third column is the new value.

Value

a transformed correlation matrix, further adjusted so that it is "positive definite"

Examples

rho <- rlkjcorr(n = 1, K = 4, eta = 1)

el <- rbind(
  c(i = 1, j = 2, r = +0.7), ## set corr between x1 and x2 as 0.7
  c(i = 3, j = 4, r = -0.5)  ## set corr between x3 and x4 as -0.5
)

out <- transform_rho(rho, el)
#> iterations: 14
#> converged: TRUE

round(rho, 2)
#>      [,1]  [,2]  [,3]  [,4]
#> [1,] 1.00  0.10  0.24  0.03
#> [2,] 0.10  1.00  0.24 -0.80
#> [3,] 0.24  0.24  1.00 -0.16
#> [4,] 0.03 -0.80 -0.16  1.00
round(out, 2)
#>       [,1]  [,2]  [,3]  [,4]
#> [1,]  1.00  0.64  0.22 -0.02
#> [2,]  0.64  1.00  0.27 -0.73
#> [3,]  0.22  0.27  1.00 -0.48
#> [4,] -0.02 -0.73 -0.48  1.00